pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-ai"
version = "0.1.0"
description = "Azure Copilot / DevOps AI assistant with a FastAPI backend and optional Streamlit UI."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Bas Franken", email = "basfrankenn@gmail.com" }]
keywords = ["devops", "fastapi", "streamlit", "azure", "ai", "llm"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Internet"
]
dependencies = [
  "Fastapi",
  "fastmcp",
  "uvicorn[standard]>=0.30,<1.0",
  "httpx>=0.27.0",
  "pydantic>=2.7,<3.0",
  "pydantic-settings>=2",
  "aiofiles",
  "python-dotenv",
  "streamlit",
  "openai>=1.0.0",
  "prometheus-client",
  "structlog",
  "PyJWT",
  "argon2-cffi",
  "asyncpg>=0.29",
  "redis",
  "cryptography",
  "numpy",
  "Pygments",
  "reportlab",
  "discord.py>=2.3.2",
  "azure-identity",
  "azure-keyvault-secrets",
  "azure-storage-blob",
  "azure-storage-file-share",
  "azure-devops",
  "azure-mgmt-resource",
  "azure-mgmt-storage",
  "azure-mgmt-network",
  "azure-mgmt-web",
  "azure-mgmt-containerregistry",
  "azure-mgmt-containerservice",
  "azure-mgmt-compute",
  "azure-mgmt-authorization",
  "azure-mgmt-sql",
  "azure-mgmt-keyvault",
  "azure-mgmt-cosmosdb",
  "azure-mgmt-redis",
  "azure-mgmt-applicationinsights",
  "azure-mgmt-loganalytics",
  "azure-mgmt-msi",
  "azure-mgmt-privatedns",
  "azure-mgmt-eventhub",
  "azure-mgmt-apimanagement",
  "azure-mgmt-datafactory",
  "azure-mgmt-synapse",
  "azure-mgmt-cognitiveservices",
  "azure-mgmt-frontdoor",
  "azure-mgmt-trafficmanager",
  "azure-mgmt-recoveryservicesbackup",
  "azure-mgmt-recoveryservices",
  "azure-mgmt-resourcegraph",
  "azure-mgmt-subscription",
  "azure-mgmt-costmanagement",
  "azure-monitor-opentelemetry>=1.7.0",
  "azure.mgmt.consumption",
  "azure.mgmt.monitor",
  "azure.mgmt.network",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-instrumentation>=0.48b0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-instrumentation-asgi>=0.48b0",
  "opentelemetry-instrumentation-httpx>=0.48b0",
  "opentelemetry-instrumentation-asyncpg>=0.48b0",
  "opentelemetry-instrumentation-psycopg2>=0.48b0",
  "opentelemetry-instrumentation-requests>=0.48b0",
  "opentelemetry-instrumentation-redis>=0.48b0",
  "mcp>=1.0.0",
  "psycopg[binary]>=3.2",
  "psycopg_pool",
  "jsonschema",
  "scikit-learn",
  "chromadb>=0.4.0",
]

[project.urls]
Repository = "https://github.com/b-franken/AzureDeployment_Ai"

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "types-requests",
  "types-PyYAML",
  "coverage>=7.6.0",
  "build>=1.2.1",
  "wheel>=0.43.0",
  "pip>=24.0",
  "scipy",
]
nlp = [
  "torch",
  "sentence-transformers"
]

[project.scripts]
devops-ai = "app.api.__main__:main"
devops-ai-mcp = "app.mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
target-version = ["py312"]
line-length = 100
exclude = """
/(
  \\.venv
 |venv
 |build
 |dist
 |__pycache__
)/
"""

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E203"]

[tool.ruff.lint.isort]
known-first-party = ["app", "ai", "bot", "kb", "memory", "tools", "workflows"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = false
strict = true
pretty = true
show_error_codes = true
exclude = "(build|dist|\\.venv|venv|__pycache__)"
files = ["src"]
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

[tool.pytest.ini_options]
minversion = "8.0"
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra --cov=src --cov-report=xml --cov-report=term-missing --junitxml=reports/junit.xml"

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["*/__init__.py", "*/tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80
